Rename dpex_exceptions.py to exceptions.py
