+++ numba_dpex/core/exceptions.py
