+++ numba_dpex/tests/kernel_tests/test_kernel_has_return_value_error.py
@@ -29,7 +29,7 @@ def sig(request):
 def test_return(filter_str, sig):
     a = np.array(np.random.random(122), np.int32)
 
-    with pytest.raises(dpex.core.dpex_exceptions.KernelHasReturnValueError):
+    with pytest.raises(dpex.core.exceptions.KernelHasReturnValueError):
         kernel = dpex.kernel(sig)(f)
 
         device = dpctl.SyclDevice(filter_str)
